Create CODE_OF_CONDUCT.md
A code of conduct is like a referee in a sports game, it helps to ensure that everyone plays fair and has a good time. By having a referee in the form of a code of conduct, we can create an awesome where everyone is treated with respect and encouraged to work together as a team. Plus it makes the project look more professional 🤣.